clipboard-apis/async-svg-unsanitized-read.https.html

<!doctype html>
<meta charset="utf-8">
<title>
  Async Clipboard unsanitized read image/svg+xml test.
</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/user-activation.js"></script>
<script>
'use strict';
// This function removes extra spaces between tags in svg. For example, the
// following svg: "<svg> <g> </g>     </svg>" would turn into this
//           svg: "<svg> <g> </g> </svg>"
// We remove the extra spaces because in svg they are considered equivalent,
// but when we are comparing for equality the spaces make a difference.
function reformatSvg(svgString) {
  const reformattedString = svgString.replace(/\>\s*\</g, '> <');
  return reformattedString;
}

const svg_with_styles =
  `<svg> <style>.foo {fill: green;}</style>
    <circle class=\"foo\" cx=\"50\" cy=\"50\" r=\"40\" /> </svg>`;
promise_test(async t => {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');
  const blobInput = new Blob([svg_with_styles], {type: 'image/svg+xml'});
  const clipboardItem = new ClipboardItem({'image/svg+xml': blobInput});
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItem]);
  await waitForUserActivation();
  const clipboardItems =
      await navigator.clipboard.read({unsanitized: ['image/svg+xml']});

  const svg = clipboardItems[0];
  assert_equals(svg.types.length, 1);
  assert_equals(svg.types[0], 'image/svg+xml');

  const blobOutput = await svg.getType('image/svg+xml');
  assert_equals(blobOutput.type, 'image/svg+xml');

  const blobText = await (new Response(blobOutput)).text();

  const outputSvg = reformatSvg(blobText);
  const inputSvg = reformatSvg(svg_with_styles);
  assert_equals(outputSvg, inputSvg);
}, 'Verify unsanitized read with styles given image/svg+xml: '
    + svg_with_styles);
</script>
